move PSE store migration to app/upgrade/v7 handler
